checked until assignment
